refractor: quizz controller and question controller
