+++ src/Form/FormQuestion.php
@@ -8,7 +8,7 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 
-class formQuestion extends AbstractType
+class FormQuestion extends AbstractType
 {
 
     public function buildForm(FormBuilderInterface $builder, array $options)
