+++ src/Form/FormQuestionc.php
@@ -7,7 +7,7 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class formQuestionc extends AbstractType
+class FormQuestionc extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options = []) : void
     {
